requirements.txt
py-wallpaper